refactor: UserRepository를 의존하게 DIP를 패키지에도 적용한다.
- domain 영역에 User와 UserRepository를 두고, 구현체들이 UserRepository를 의존하게 변경하여 DIP를 적용함
- 재사용성을 높일 수 있을것으로 기대하기 때문에 작업함
