+++ src/main/java/codesquad/web/domain/User.java
@@ -1,4 +1,4 @@
-package codesquad.web.model;
+package codesquad.web.domain;
 
 public class User {
