+++ src/main/java/codesquad/web/handler/LoginRequestHandler.java
@@ -6,8 +6,8 @@ import codesquad.was.http.Cookie;
 import codesquad.was.http.HttpRequest;
 import codesquad.was.http.HttpResponse;
 import codesquad.was.http.HttpSession;
-import codesquad.web.model.User;
-import codesquad.web.repository.UserRepository;
+import codesquad.web.domain.User;
+import codesquad.web.domain.UserRepository;
 import java.io.IOException;
 import java.util.Objects;
 import java.util.UUID;
