+++ src/main/java/codesquad/web/handler/SignUpRequestHandler.java
@@ -4,8 +4,8 @@ import codesquad.was.AbstractRequestHandler;
 import codesquad.was.exception.BadRequestException;
 import codesquad.was.http.HttpRequest;
 import codesquad.was.http.HttpResponse;
-import codesquad.web.model.User;
-import codesquad.web.repository.UserRepository;
+import codesquad.web.domain.User;
+import codesquad.web.domain.UserRepository;
 import java.io.IOException;
 import java.util.Objects;
 import org.slf4j.Logger;
