+++ src/main/java/codesquad/web/handler/UserRequestHandler.java
@@ -6,8 +6,8 @@ import codesquad.was.http.HttpRequest;
 import codesquad.was.http.HttpResponse;
 import codesquad.was.http.HttpSession;
 import codesquad.was.http.type.MimeType;
-import codesquad.web.model.User;
-import codesquad.web.repository.UserRepository;
+import codesquad.web.domain.User;
+import codesquad.web.domain.UserRepository;
 import codesquad.web.snippet.ResourceSnippetBuilder;
 import codesquad.web.snippet.Snippet;
 import codesquad.web.snippet.SnippetBuilder;
