+++ src/test/java/codesquad/web/handler/RequestHandlerTest.java
@@ -5,8 +5,8 @@ import codesquad.was.http.Headers;
 import codesquad.was.http.HttpRequest;
 import codesquad.was.http.HttpResponse;
 import codesquad.was.http.type.HeaderType;
-import codesquad.web.repository.JdbcUserRepository;
-import codesquad.web.repository.UserRepository;
+import codesquad.web.infrastructure.JdbcUserRepository;
+import codesquad.web.domain.UserRepository;
 import java.io.ByteArrayInputStream;
 import java.io.IOException;
 import java.io.InputStream;
